refactor: (#103) 비밀번호 초기화 변경사항
