+++ simtong-presentation/src/main/kotlin/team/comit/simtong/user/WebUserAdapter.kt
@@ -12,21 +12,18 @@ import team.comit.simtong.domain.auth.dto.TokenResponse
 import team.comit.simtong.domain.user.dto.ChangeEmailRequest
 import team.comit.simtong.domain.user.dto.ChangeNicknameRequest
 import team.comit.simtong.domain.user.dto.ChangeProfileImageRequest
-import team.comit.simtong.domain.user.dto.ResetPasswordRequest
 import team.comit.simtong.domain.user.dto.SignInRequest
 import team.comit.simtong.domain.user.dto.SignUpRequest
 import team.comit.simtong.domain.user.dto.UserInfoResponse
 import team.comit.simtong.domain.user.usecase.ChangeEmailUseCase
 import team.comit.simtong.domain.user.usecase.ChangeNicknameUseCase
 import team.comit.simtong.domain.user.usecase.ChangeProfileImageUseCase
-import team.comit.simtong.domain.user.usecase.ResetPasswordUseCase
 import team.comit.simtong.domain.user.usecase.SignInUseCase
 import team.comit.simtong.domain.user.usecase.SignUpUseCase
 import team.comit.simtong.domain.user.usecase.UserInfoUseCase
 import team.comit.simtong.user.dto.request.WebChangeEmailRequest
 import team.comit.simtong.user.dto.request.WebChangeNicknameRequest
 import team.comit.simtong.user.dto.request.WebChangeProfileImageRequest
-import team.comit.simtong.user.dto.request.WebResetPasswordRequest
 import team.comit.simtong.user.dto.request.WebSignInRequest
 import team.comit.simtong.user.dto.request.WebSignUpRequest
 import javax.validation.Valid
@@ -48,8 +45,7 @@ class WebUserAdapter(
     private val getInfoUseCase: UserInfoUseCase,
     private val changeEmailUseCase: ChangeEmailUseCase,
     private val changeNicknameUseCase: ChangeNicknameUseCase,
-    private val changeProfileImageUseCase: ChangeProfileImageUseCase,
-    private val resetPasswordUseCase: ResetPasswordUseCase
+    private val changeProfileImageUseCase: ChangeProfileImageUseCase
 ) {
 
     @ResponseStatus(HttpStatus.CREATED)
@@ -88,18 +84,6 @@ class WebUserAdapter(
         changeNicknameUseCase.execute(ChangeNicknameRequest(
             nickname = request.nickname
         ))
-    }
-    
-    @PutMapping("/password/initialization")
-    @ResponseStatus(HttpStatus.NO_CONTENT)
-    fun resetPassword(@Valid @RequestBody request: WebResetPasswordRequest) {
-        resetPasswordUseCase.execute(
-            ResetPasswordRequest(
-                email = request.email,
-                employeeNumber = request.employeeNumber,
-                newPassword = request.newPassword
-            )
-        )
     }
 
     @PutMapping("/email")
